chore: linter fixes, moving things
