+++ suite-native/assets/tsconfig.json
@@ -15,6 +15,7 @@
             "path": "../../suite-common/wallet-utils"
         },
         { "path": "../atoms" },
+        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../module-settings" },
         { "path": "../navigation" },
