+++ suite-native/fiat-rates/package.json
@@ -16,7 +16,10 @@
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-config": "workspace:*",
+        "@suite-common/wallet-config": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
+        "@suite-common/wallet-utils": "workspace:*",
         "@trezor/connect": "workspace:9.0.7",
         "date-fns": "^2.29.3",
         "proxy-memoize": "^2.0.2"
