+++ suite-native/fiat-rates/src/index.ts
@@ -3,3 +3,4 @@ export * from './fiatRatesActions';
 export * from './fiatRatesSelectors';
 export * from './fiatRatesThunks';
 export * from './fiatRatesMiddleware';
+export * from './types';
