+++ suite-native/fiat-rates/tsconfig.json
@@ -12,7 +12,16 @@
             "path": "../../suite-common/suite-config"
         },
         {
+            "path": "../../suite-common/wallet-config"
+        },
+        {
+            "path": "../../suite-common/wallet-core"
+        },
+        {
             "path": "../../suite-common/wallet-types"
+        },
+        {
+            "path": "../../suite-common/wallet-utils"
         },
         { "path": "../../packages/connect" }
     ]
