+++ suite-native/formatters/tsconfig.json
@@ -22,6 +22,7 @@
         },
         { "path": "../atoms" },
         { "path": "../ethereum-tokens" },
+        { "path": "../fiat-rates" },
         { "path": "../module-settings" },
         {
             "path": "../../packages/blockchain-link"
