+++ suite-native/logger/tsconfig.json
@@ -9,6 +9,7 @@
         {
             "path": "../../suite-common/wallet-core"
         },
+        { "path": "../fiat-rates" },
         { "path": "../module-settings" }
     ]
 }
