+++ suite-native/module-accounts-import/package.json
@@ -29,6 +29,7 @@
         "@suite-native/atoms": "workspace:*",
         "@suite-native/config": "workspace:*",
         "@suite-native/ethereum-tokens": "workspace:*",
+        "@suite-native/fiat-rates": "workspace:*",
         "@suite-native/formatters": "workspace:*",
         "@suite-native/forms": "workspace:*",
         "@suite-native/module-devices": "workspace:*",
