+++ suite-native/module-accounts-import/tsconfig.json
@@ -25,6 +25,7 @@
         { "path": "../atoms" },
         { "path": "../config" },
         { "path": "../ethereum-tokens" },
+        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../forms" },
         { "path": "../module-devices" },
