+++ suite-native/state/src/store.ts
@@ -4,10 +4,10 @@ import createDebugger from 'redux-flipper';
 import { prepareFiatRatesMiddleware } from '@suite-native/fiat-rates';
 import { logsMiddleware } from '@suite-common/logger';
 import { logsMiddleware as nativeLogsMiddleware } from '@suite-native/logger';
+import { prepareBlockchainMiddleware } from '@suite-common/wallet-core';
 
 import { extraDependencies } from './extraDependencies';
 import { prepareRootReducers } from './reducers';
-import { prepareBlockchainMiddleware } from '@suite-common/wallet-core';
 
 const middlewares: Middleware[] = [
     nativeLogsMiddleware,
