+++ suite-native/transactions/tsconfig.json
@@ -25,6 +25,7 @@
         },
         { "path": "../atoms" },
         { "path": "../ethereum-tokens" },
+        { "path": "../fiat-rates" },
         { "path": "../formatters" },
         { "path": "../helpers" },
         { "path": "../module-settings" },
